# Avoids reparses in the signature help