--- a/packages/language-support/src/helpers.ts
+++ b/packages/language-support/src/helpers.ts
@@ -56,7 +56,7 @@ export function findStopNode(root: StatementsContext) {
   return current;
 }
 
-export function findMostApproximateNode(
+export function findNearestNodeStartingBeforeOrAt(
   current: ParseTree,
   position: number,
 ): EnrichedParseTree | undefined {
@@ -80,7 +80,30 @@ export function findMostApproximateNode(
       while (index < children.length) {
         const child = children[index];
 
-        const maybeResult = findMostApproximateNode(child, position);
+        /*
+          We'll try to pick the most specific node whose position is lower or equal to where the caret is
+
+              i.e. for CALL procedure(true
+                                   ^
+          we would have a parsing tree akin to this one:
+
+                     procedure: CALL ( expression
+                                           |
+                                    boolean literal
+                                           |
+                                          true
+ 
+
+          we would return the literal true.
+
+          If the caret was at a space, we would pick what's inmediately before the space:
+
+             CALL procedure(true,  
+                                   ^
+          we would pick the COMMA element. This is because spaces don't appear in the 
+          parse tree, so we want to pick whatever rule is closest to them
+        */
+        const maybeResult = findNearestNodeStartingBeforeOrAt(child, position);
         if (maybeResult) {
           result = maybeResult;
         }