## What
Adds a `caretPosition` parameter to the signature help to avoid reparses

## Why
Because that way if we have a written query and we move between function arguments, we don't need to chop the query until the caret position and reparse. We can just send the full query and extract the information we need to know in what argument are we from the caret position. 

Until now for the query:

```
MATCH (n) WHERE apoc.coll.combinations(coll, "foo", bar
                                                       ^
```

if we moved the caret `^` to an argument before and retriger signature help:

```
MATCH (n) WHERE apoc.coll.combinations(coll, "foo", bar
                                          ^
```

we would send `MATCH (n) WHERE apoc.coll.combinations(` to the signature help underlying methods, causing a reparse cause only the full `MATCH (n) WHERE apoc.coll.combinations(coll, "foo", bar` was in the parsing cache.

From now on, we would send `MATCH (n) WHERE apoc.coll.combinations(coll, "foo", bar` with the corresponding caret position, which avoids reparses because the full query is found in the cache. The trade off is we need to do more computations to calculate where we are in the parsing tree for that caret position.


## Benchmarks

Before:
```
movies - signature help x 17.08 ops/sec ±2.19% (39 runs sampled)
tictactoe - signature help x 32.80 ops/sec ±1.23% (59 runs sampled)
pokemon - signature help x 1.35 ops/sec ±2.51% (13 runs sampled)
```

After:

```
movies - signature help x 24.17 ops/sec ±4.52% (17 runs sampled)
tictactoe - signature help x 58.77 ops/sec ±4.55% (21 runs sampled)
pokemon - signature help x 2.03 ops/sec ±2.07% (11 runs sampled)
```